requirements.txt
Flask>=3.1.2
pdfplumber
python-dotenv
torch 
transformers
flask-mail


